requirements.txt
aspose-zip
gradio